Reorganize components and stories
