+++ client/src/app/components/pages/SendPage.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import SendBeginScreen from "./screens/send/SendBeginScreen";
-import SendCompleteScreen from "./screens/send/SendCompleteScreen";
-import SendInstructionsScreen from "./screens/send/SendInstructionsScreen";
-import SendProgressScreen from "./screens/send/SendProgressScreen";
+import SendBeginScreen from "../screens/send/SendBeginScreen";
+import SendCompleteScreen from "../screens/send/SendCompleteScreen";
+import SendInstructionsScreen from "../screens/send/SendInstructionsScreen";
+import SendProgressScreen from "../screens/send/SendProgressScreen";
 
 type Props = {
   step: "BEGIN" | "INSTRUCTIONS" | "PROGRESS" | "DONE";
@@ -19,4 +19,9 @@ export function SendPageContent(props: Props) {
     case "DONE":
       return <SendCompleteScreen />;
   }
+}
+
+export default function SendPage() {
+  // TODO
+  return <SendPageContent step="BEGIN" />;
 }
