+++ client/src/stories/screens/send/SendInstructionsScreen.stories.tsx
@@ -1,8 +1,8 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
-import AppTemplate from "../app/components/AppTemplate";
-import { SendInstructionsScreenContent } from "../app/components/screens/send/SendInstructionsScreen";
+import AppTemplate from "../../../app/components/AppTemplate";
+import { SendInstructionsScreenContent } from "../../../app/components/screens/send/SendInstructionsScreen";
 
 export default {
   title: "screens/SendInstructionsScreen",
